backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
jinja2
docker
websockets
aiofiles
certbot
certbot-dns-cloudflare